backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Image processing
Pillow==10.1.0
exifread==3.0.0

# AI/ML
tensorflow==2.15.0
tensorflow-hub==0.15.0
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# CORS
python-multipart==0.0.6